requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
docutils==0.21.2
filetype==1.2.0
future==1.0.0
idna==3.10
iso8601==2.1.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kivy-garden.graph==0.4.0
kivy_deps.sdl2==0.8.0
paho-mqtt==2.1.0
pillow==10.4.0
Pygments==2.19.1
pypiwin32==223
pyserial==3.5
pywin32==308
PyYAML==6.0.2
requests==2.32.3
ulid-py==1.1.0
urllib3==2.3.0
websockets==14.1
